Clear joined channels on connect
Clears joined channels when calling Connect()
Issue occurs when the library gets disconnected from a network issue and a new connection is established. Because the library didn't know it was disconnected it still thinks it's joined to the channels it was previously in. Unless you use the force join flag when calling the join channel function all calls to join channels you were previously in will be ignored because it thinks you're still in them.