Fixed joining Channel on IClient Reconnect()
While TwitchClient.Reconnect() currently joins all channels, it wont join all channel if the IClient only was reconencted. I moved adding the currently joined channel to the joining queue from Reconnect() to _client_OnReconnected() since the Reconnect() Methode calls the _client.Reconnect() methode which calls the _client_OnReconnected Event. To ensure that the BotAccount doesnt get joined twice, i check if it is not him and then add to the queue.

To get this work i need to remove the _joinedChannelManager.Clear() from _client_OnDisconnected since it gets deleted on the TwitchClient.Disconnect() Methode but stays if only the IClient gets disconnected to ensure it can be queued again after the reconnect.